requirements.txt
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
gcloud==0.18.3
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.4
jwcrypto==1.5.6
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
oauth2client==4.1.3
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==5.28.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
pyparsing==3.1.4
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-jwt==4.1.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
scipy==1.14.1
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.2.103
ultralytics-thop==2.0.8
urllib3==1.26.20
